Task/update to python3 (#1)
* moved files

* add pylint files

* removing tabs

* removal of tabs

* lint changes

* added version

* removed tabs

* changes
